Scripts/requirements.alt.txt
transparent_background
tqdm
Pillow